feat: next.config.ts 파일 추가 및 Vanilla Extract 플러그인 설정
