+++ next.config.ts
@@ -1,0 +1,48 @@
+import { createVanillaExtractPlugin } from "@vanilla-extract/next-plugin";
+import type { NextConfig } from "next";
+
+const withVanillaExtract = createVanillaExtractPlugin();
+
+const nextConfig: NextConfig = {
+  basePath: "",
+  pageExtensions: ["js", "jsx", "mdx", "ts", "tsx"],
+  images: {
+    formats: ["image/avif", "image/webp"],
+  },
+  compiler: {
+    styledComponents: true,
+  },
+  async headers() {
+    const cspHeader = `
+      default-src 'self';
+      script-src 'self' 'unsafe-eval' 'unsafe-inline';
+      style-src 'self' 'unsafe-inline' https://fonts.googleapis.com;
+      img-src 'self' blob: data:;
+      font-src 'self' https://fonts.gstatic.com;
+      object-src 'none';
+      base-uri 'self';
+      form-action 'self';
+      frame-ancestors 'none';
+      block-all-mixed-content;
+      upgrade-insecure-requests;
+    `;
+
+    return [
+      {
+        source: "/:path*",
+        headers: [
+          {
+            key: "Content-Security-Policy",
+            value: cspHeader.replace(/\s{2,}/g, " ").trim(),
+          },
+          {
+            key: "X-Requested-With",
+            value: "XMLHttpRequest",
+          },
+        ],
+      },
+    ];
+  },
+};
+
+export default withVanillaExtract(nextConfig);
